workspace/web-workspace/system-viewer/backend/requirements.txt
djangorestframework==3.11.0
drf-url-filters==0.5.1